requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.3
fonttools==4.61.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
numba==0.64.0
numpy==2.2.6
opencv-python==4.13.0.92
optiland==0.5.9
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
tabulate==0.10.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vtk==9.6.0
Werkzeug==3.1.6
